class spectrum, right channel path, change colours
oscillator added to pluginprocessor to make some annoying sine wave tests

below is a bunch of troubleshooting i tried when trying to paint() both channel paths to the spectrum analyzer. the FTT generator puts the blocks into a queue but this completely shits the bed on small buffer sizes. my audio interface was set to 48kHz and 128 buffer size, which was just too small to use. anything over 500 buffer size seems to work okay, so I'm guessing its a multiple math problem (496?). increasing the buffer size adds latency, so there's a big trade-off for being able to constantly repaint the GUI window to show the buffered blocks of audio data and having it work at all

pathproducer and the added classes work but when trying to repaint() both paths in the timercallback its just too much for the GUI to handle and will either crash or freeze, even without playing anything through it

startTimerHz(60); can be played around with but doesn't really help for performance

tried adding/changing the orders which will either do nothing or hard crash the memory...
enum FFTOrder
{
    order2048 = 11,
    order4096 = 12,
    order8192 = 13
};

changing the binsize will just cause the rendering of the lines to be whacked since it's not drawing the correct hz

at the moment it feels like the only thing to do is to run the spectrum analyzer in mono only, commenting out the right path from timerCallback() and not drawing the strokepath or setting the rightpath in ResponseCurveComponent::paint, or just whack up the buffer size

add fonts to projucer file explorer to add them as a binary resource, then add to cmake with juce_add_binary_data(), then add a pointer to the typeface and set the font:

    const juce::Typeface::Ptr typeface = juce::Typeface::createSystemTypefaceFor(BinaryData::Monomaniac_ttf, BinaryData::Monomaniac_ttfSize);
    g.setFont(juce::Font(typeface).withHeight(11.0f));